Major change on the cdc interface and changes for extracting deltas at high database loads.
No state is held in cdc.mysql-binlog. Agent is initialized from the
API-consumer. Likewise, a fn can now be specified to handle parsed
binlog-events, eg. push them on a queue or simply print them.
Examples for using cdc.mysql-binlog are shown at the bottom of the file.

Bypassing strange INotify behaviour:
When removing a inotify watch and immediately adding another to a new
file, the new watch will have the same id of the former watch. But
after removing the old watch, a remove event will be send with the id
of the old(and new) watch. JNotify will then remove the watch from its
internal c-java mapping of watch ids, and therefor remove the new
watch. This happens only under certain circumstances (e.g. high
database/cpu loads). The fix is to create the new watch first and
remove the old watch after that with the effect that new and old will
have different ids. Were not expecting any events sent for the old
logfile anyways.